requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
phonenumberslite==8.13.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
